# Swager/krish.postman_collection.json-OpenApi3Json.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "krish",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5555/api",
      "variables": {}
    }
  ],
  "paths": {
...
    "/passwordupdate": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "passwordupdate",
        "operationId": "passwordupdate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/passwordupdaterequest"
                  },
                  {
                    "example": {
                      "email": "abc@gmail.com",
                      "password": "bbbbb",
                      "confirmpassword": "bbbbb"
                    }
                  }
                ]
              },
              "example": {
                "email": "abc@gmail.com",
                "password": "bbbbb",
                "confirmpassword": "bbbbb"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    }
  },
  "components": {
    "schemas": {
...
      "passwordupdaterequest": {
        "title": "passwordupdaterequest",
        "required": [
          "email",
          "password",
          "confirmpassword"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmpassword": {
            "type": "string"
          }
        },
        "example": {
          "email": "abc@gmail.com",
          "password": "bbbbb",
          "confirmpassword": "bbbbb"
        }
      }
    }
  },
...
}